+++ db/migrate/20220324105755_add_sti_type_to_sites.rb
@@ -1,0 +1,8 @@
+# frozen_string_literal: true
+
+class AddStiTypeToSites < ActiveRecord::Migration[7.0]
+  def change
+    add_column :folio_sites, :type, :string
+    add_index :folio_sites, :type
+  end
+end